Tweaks to snow course imports.

+++ R/synchronize.R
@@ -117,6 +117,7 @@ synchronize <- function(con = hydrometConnect(silent = TRUE), timeseries_id = "a
       }
       if (source_fx == "downloadSnowCourse") {
         args_list[["snowCon"]] <- snowCon
+        args_list[["hydroCon"]] <- con
       }
       inRemote <- do.call(source_fx, args_list) #Get the data using the args_list
       inRemote <- inRemote[!is.na(inRemote$value) , ]
@@ -135,7 +136,7 @@ synchronize <- function(con = hydrometConnect(silent = TRUE), timeseries_id = "a
             }
           }
         } else if (category == "discrete") {
-          inDB <- DBI::dbGetQuery(con, paste0("SELECT target_datetime, datetime, value, sample_class FROM measurements_discrete WHERE timeseries_id = ", tsid, " AND datetime >= '", min(inRemote$datetime),"';"))
+          inDB <- DBI::dbGetQuery(con, paste0("SELECT target_datetime, datetime, value, note, sample_class FROM measurements_discrete WHERE timeseries_id = ", tsid, " AND datetime >= '", min(inRemote$datetime),"';"))
         }
         
         if (nrow(inDB) > 0) {
@@ -143,42 +144,47 @@ synchronize <- function(con = hydrometConnect(silent = TRUE), timeseries_id = "a
             inDB <- inDB[inDB$datetime >= min(inRemote$datetime) , ]
           }
           
-          #order both timeseries to compare them
-          inDB <- inDB[order(inDB$datetime) , ]
-          inRemote <- inRemote[order(inRemote$datetime) , ]
-          
-          # Create a unique datetime key for both data frames
-          if (category == "continuous") {
-            inRemote$key <- paste(inRemote$datetime, inRemote$value, inRemote$grade, inRemote$approval, sep = "|")
-            inDB$key <- paste(inDB$datetime, inDB$value, inDB$grade, inDB$approval, sep = "|")
-          } else if (category == "discrete") {
-            inRemote$key <- paste(inRemote$target_datetime, inRemote$datetime, inRemote$value, inRemote$sample_class, sep = "|")
-            inDB$key <- paste(inDB$target_datetime, inDB$datetime, inDB$value, inDB$sample_class, sep = "|")
-          }
-          
-          # Check for mismatches using set operations
-          mismatch_keys <- inDB$key[!(inDB$key %in% inRemote$key)]
-          
-          # Check where the discrepancy is in both data frames
-          if (length(mismatch_keys) > 0) {
+          if (min(inRemote$datetime) < min(inDB$datetime)) { #if TRUE means that the remote has older data than the DB, so immediately declare mismatch = TRUE.
             mismatch <- TRUE
-            datetime_remote <- inRemote[inRemote$key %in% mismatch_keys, "datetime"]
-            if (length(datetime_remote) == 0) { #Means that the remote data doesn't exist or is different for the mismatch point. Delete from that point on in the DB
-              datetime_db <- inDB[inDB$key %in% mismatch_keys, "datetime"]
-              datetime_db <- min(datetime_db)
-              index <- which(inDB$datetime == datetime_db)
-              if (index > 1) {
-                datetime <- inDB[index - 1, "datetime"]
-              } else if (index == 1) {
-                datetime <- inDB[index, "datetime"]
+            datetime <- min(inRemote$datetime)
+          } else {
+            #order both timeseries to compare them
+            inDB <- inDB[order(inDB$datetime) , ]
+            inRemote <- inRemote[order(inRemote$datetime) , ]
+            
+            # Create a unique datetime key for both data frames
+            if (category == "continuous") {
+              inRemote$key <- paste(inRemote$datetime, inRemote$value, inRemote$grade, inRemote$approval, sep = "|")
+              inDB$key <- paste(inDB$datetime, inDB$value, inDB$grade, inDB$approval, sep = "|")
+            } else if (category == "discrete") {
+              inRemote$key <- paste(inRemote$target_datetime, inRemote$datetime, inRemote$value, inDB$note, inRemote$sample_class, sep = "|")
+              inDB$key <- paste(inDB$target_datetime, inDB$datetime, inDB$value, inDB$note, inDB$sample_class, sep = "|")
+            }
+            
+            # Check for mismatches using set operations
+            mismatch_keys <- inDB$key[!(inDB$key %in% inRemote$key)]
+            
+            # Check where the discrepancy is in both data frames
+            if (length(mismatch_keys) > 0) {
+              mismatch <- TRUE
+              datetime_remote <- inRemote[inRemote$key %in% mismatch_keys, "datetime"]
+              if (length(datetime_remote) == 0) { #Means that the remote data doesn't exist or is different for the mismatch point. Delete from that point on in the DB
+                datetime_db <- inDB[inDB$key %in% mismatch_keys, "datetime"]
+                datetime_db <- min(datetime_db)
+                index <- which(inDB$datetime == datetime_db)
+                if (index > 1) {
+                  datetime <- inDB[index - 1, "datetime"]
+                } else if (index == 1) {
+                  datetime <- inDB[index, "datetime"]
+                }
+              } else {
+                datetime <- min(datetime_remote)
               }
             } else {
-              datetime <- min(datetime_remote)
+              mismatch <- FALSE
             }
-          } else {
-            mismatch <- FALSE
+            inRemote$key <- NULL
           }
-          inRemote$key <- NULL
         } else {
           mismatch <- TRUE
           datetime <- min(inRemote$datetime)
